Autonomous strafing patch v2.0
